bytebooks-api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
httpx>=0.24.0
passlib[bcrypt]
bcrypt<4.0
python-jose[cryptography]
pydantic[email]
psycopg2-binary
